pyproject.toml
[tool.poetry]
name = "langchain-poc"
version = "0.1.0"
description = ""
authors = ["Scott Yang <yisyang@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}, {include = "serverless"}]

[tool.poetry.dependencies]
python = "~=3.8.6"
langchain = "0.0.141"
openai = "~=0.27"
fastapi = "~=0.95"
click = "~=8.1.3"
wikipedia = "^1.4.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
uvicorn = "^0.21.1"
httpx = "^0.24.0"
pytest = "^7.3.1"
pytest-cov = "^4.0.0"

[tool.poetry.scripts]
start = "src.api.main:main"
cli = "src.cli.main:cli"
hello = "src.cli.main:say_hello"
agent = "src.cli.main:ask_agent"
test = "pytest:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
